requirements.txt
requests==2.4
python-twitter==2.0
python-instagram==1.1.3
gspread
